requirements.txt
reportlab>=4.0
num2words>=0.5.12
python-dotenv>=1.0.0
flask>=3.0
arabic-reshaper>=3.0
python-bidi>=0.4.2
gunicorn>=21.2.0
